res11.27

+++ controllers/contacts/removeContact.js
@@ -1,8 +1,9 @@
 const {Contact} = require('../../models/contact');
 
 const removeContact = async (req, res) => {
+   const { _id } = req.user;
    const {contactId} = req.params; 
-   const removeContact = await Contact.findByIdAndRemove(contactId);
+   const removeContact = await Contact.findByIdAndRemove({owner: _id, _id: contactId});
    if(removeContact) {
    res.status(200).json({ message: 'contact deleted' });   
    } else {
